Detection of online censorship, estimation of platforms’ social biases/constraints, and derivation of responsibility
With the more popular use of the N-word since Elon Musk’s takeover of the
Twitter company we came to the idea of figuring out whose responsibility the use
of koˇcljivih(delicate, socially sensitive) (taboo, forbidden, auto-censorship,
social barrier) terms is; the company’s, the user’s, state legislator’s? And to
what extent? ‘ We will attempt to find more of such terms